Fix SSR rendering issues for chatbox, dark mode, TOS, and locale
For non-default configurations, ie. language other than English, dark mode, chat box open, and TOS not accepted, nextJS was complaining about server and client state mismatch. This is because the server would render the site in the default configuration and the client would switch it to a different configuration.

To bring the server and client rendered versions in sync, the rendering process should either a. not rely on client-specific mechanisms such as localStorage or b. not render the component on the server side at all.

For dark mode setting and locale, used approach a., since those are required for the layout. Using approach b for these would mean that we lose all the benefits of SSR.

For chat box position and TOS timestamp, used approach b. since those can be rendered after the page has been hydrated on the client side.