ARC-2248: Update docker image tag for Labeeb to include a unique ID on every run
Currently, the image tag is equal to the git commit SHA. This means that if the environment variables are changed and the job is re-run, the image tag stays the same. This causes Azure to not pick up the latest image. To address this, we need to add another piece of information to the image tag, such as timestamp or run ID.