ARC-2289: Scroll to bottom after images/videos are done loading
The issue with this bug was that the `ScrollToBottom` code would execute as soon as the image container was added to the message list. The problem with this is that the after the image actually loads, the container may resize depending on the image size. So the initial height calculated by the `ScrollToBottom` component was outdated and didn't scroll down far enough. To fix it, we're now keeping track of loading states. Images and videos are marked as "loaded" when they actually load. When loading is complete, the `ScrollToBottom` code kicks in again.