ARC-2318: Async requests failing with an error
Websockets requests were not being passed through correctly from the middleware, resulting in an exception in one of the nextjs server layers. I tried to make it work, but due to lack of documentation and an obvious way to do it, I'm not even sure if it's supported. So I bypassed `/graphql` paths from the middleware.