ARC-2303: Client re-creation causes GraphQL queries to re-run
After the migration to NextJS, we found that Next can render the layout element multiples times during the course of running the app, causing the apollo client to be re-created. The Apollo team has released an experimental package that helps with this by creating the client once and re-using it.