ARC-2461: Digest generation job may leave one or more digests in the "in_progress" state
This was happening due to a race condition in DB writes. We can't effectively do parallel writes to specific items in the blocks array because the blocks field is encrypted. So we have to update the entire array. The best option then is to write the update out later, after all generations have completed.