Fix LLM seeding creating duplicate LLM's
I saw in the dev environment that there were some duplicate llm's and multiple llm's had the wrong identifier, in particular gpt35turbo. Tracked it down to this line of code:

```
const identifier = config.data.llms.find(
                    (llm) => llm.cortexModelName === llm.cortexModelName,
```

This always returns the first llm, so identifier is always the first llm's identifier, and it's assigned to every llm.

Then the subsequent code goes and finds a bunch of missing identifiers and creates new llm's for each one.

Added tests as well.