Allow user to navigate away from Video page when operations are in progress
If the user navigates away from the Video page when a transcription or translation operation is happening, `setTranscripts` doesn't fire and as a result, user state is never updated. The transcription is lost.

With these new changes, the user is able to do other things in the app and the completion handler will still do what its supposed to do.

This will not be required once we move tasks to offline workers, but for now this does the job