Only overwrite local state from the server version on first load
We're seeing some user-state-backed inputs revert to earlier values.

The root cause is the following race condition:
1. updates have debouncing logic so they're sent to the server with a delay
2. if the value has changed since the time the state was sent to the server, the value that comes back is stale and applying it to the client copy of the state results in the value being reverted

The fix is to only apply server state to client if client state is empty.

This bug was introduced in commit 234676c166c099cbd4e0a19aed5c3a570b9afa2f. Since video translation happens offline (in the workers), the state change happens on the server and needs to be propagated to the client.

To propgate the change to the client, we're now using the refetch method and clearing out the client state. After refetch is complete, the client state will be empty, so the server state will be applied to the client.