Switch imageUtils import to dynamic import
While running in the production (Docker) context, imageUtils was throwing an error saying import statements are not supported. Switched to dynamic imports to fix this. Also, directory imports are not supported (they require some kind of build layer) when running directly in node, so made it import `index.js` explicitly.