Workspace creation does unnecessary find calls
The current workspace creation function tries to find a suffix for the name and slug that's not taken. It does this by doing find queries: one for the name for each index it tries and one for slug. This naturally doesn't scale well. e.g. if there are New Workspace 1-200 are taken, it's gonna try all 200 of them before it settles on "New Workspace 201".

So the fix is to:
1. always use "New Workspace" as the name. These don't have to be unique
2. generate a random hex slug. If it's taken, generate another one. Maximum 3 attempts, so max 3 finds per workspace creation and then it'll fail.