Workspace input should not reset after state update
Workspace input should keep a client copy and not overwrite it with state updates. Doing so introduces a race condition where the updated state from the server overwrites the client copy, causing it to potentially lose any changes that the user may have made since the last request was sent. In terms of UX, this results in input being "reset" to an older point arbitrarily during use.