Monaco editor grows indefinitely for published workspaces
When the workspace is published, we have a notice that says that the code is read-only. This means that Monaco is no longer the only child of its parent. Monaco doesn’t like this configuration and so keeps growing indefinitely. This may the same as this bug: https://github.com/react-monaco-editor/react-monaco-editor/issues/391