For LLM output in applets, handle the case where the element is already displayed
The issue here was that if the applet renders the LLM output field empty (e.g. before the full response has been received), we "process" it and replace it with OutputSandbox, but because it doesn't contain any content, it's blank. And then, when the actual LLM response comes in, it's not reprocessed. So the fix is to listen for changes to the pre and process it everytime the content changes